requirements.txt
openai
scikit-learn
dotenv
matplotlib
umap-learn
seaborn
openai-whisper
django